Review changes: Moved Visitors and Visitor Sessions storage to its own class. Removed caching of api clients.
